requirements.txt
PyYAML==6.0
pymongo==4.3.3
dnspython==2.3.0
Flask==2.3.2
blinker==1.6.2
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.3.4
flask-swagger-ui==4.11.1
Flask-RESTful==0.3.10